refactor(app-file-ops): extract new-file editing and validation modules
